+++ src/components/IdeaCard/IdeaCard.test.js
@@ -1,15 +1,29 @@
 import React from 'react';
 import { shallow, mount } from 'enzyme';
 import IdeaCard from './';
+import { Provider } from 'react-redux';
+import configureStore from 'redux-mock-store'
 import { teams } from '../../server/mockServerData'
 
 describe('IdeaCard', () => {
+	const initialState = {
+		open: false,
+		ideaContent: {
+			id: '',
+			name: '',
+			description: '',
+			team: {}
+		}
+	};
+	const mockStore = configureStore();
+	let store = mockStore(initialState);
+
 	it('renders without crashing', () => {
-		expect(shallow.bind(shallow, <IdeaCard/>)).not.toThrow();
+		expect(mount.bind(shallow, <Provider store={store}><IdeaCard /></Provider>)).not.toThrow();
 	});
 
 	it('renders as expected', () => {
-		expect(shallow(<IdeaCard/>)).toMatchSnapshot();
+		expect(mount(<Provider store={store}><IdeaCard /></Provider>)).toMatchSnapshot();
 	});
 
 	describe('given props', () => {
@@ -22,7 +36,8 @@ describe('IdeaCard', () => {
 		};
 
 		beforeEach(() => {
-			card = mount(<IdeaCard { ...props }/>);
+			store = mockStore(initialState);
+			card = mount(<Provider store={store}><IdeaCard { ...props }/></Provider>);
 		});
 		it('renders content', () => {
 			expect(card.find('.IdeaCard-Info-Name').text()).toBe(props.name);